// frontend/src/app/models/patient.doctor.view.ts
export class PatientDoctorProfileData{
    _id: Object;
	firstname: string;
	lastname: string;
	username: string;
	password: string;
	address: string;
	mobile_phone:  string;
	email:  string;
	type: string;
	d_data: Array<any>;
	img_src: string;
	status: string;
    img_profile: Buffer;
    img_path : string;
}